anemo
Struct Builder

pub struct Builder { /* private fields */ }

A builder for a Network.

Implementations

impl Builder

pub fn config(self, config: Config) -> Self

Set the Config that this network should use.

pub fn server_name<T: Into<String>>(self, server_name: T) -> Self

Set the server-name that will be used when constructing a self-signed X.509 certificate to be used in the TLS handshake.

Traditionally a server name is intended to be the DNS name that is being dialed, although since Anemo does not require parties to use DNS names, nor does it rely on a central CA, server-name is instead used to identify the network name that this Peer can connect to. In other words, The TLS handshake will only be successful if all parties use the same server-name.

pub fn alternate_server_name<T: Into<String>>(self, server_name: T) -> Self

alternate-server-name helps with network name migration. Server accepts connections from both server-name and alternate-server-name. However client only initiates connection with server-name.

pub fn private_key(self, private_key: [u8; 32]) -> Self

Set the Ed25519 Private Key that will be used to perform the TLS handshake. The corresponding Public Key will be this node’s PeerId.

pub fn outbound_request_layer<L>(self, layer: L) -> Self
where L: Layer<BoxService<Request<Bytes>, Response<Bytes>, Error>> + Send + Sync + 'static, L::Service: Service<Request<Bytes>, Response = Response<Bytes>, Error = Error> + Send + 'static, <L::Service as Service<Request<Bytes>>>::Future: Send + 'static,

Provide an optional [Layer] that will be used to wrap all outbound RPCs.

This could be helpful in providing global metrics and logging for all outbound requests

pub fn start<T>(self, service: T) -> Result<Network>
where T: Clone + Send + 'static + Service<Request<Bytes>, Response = Response<Bytes>, Error = Infallible>, <T as Service<Request<Bytes>>>::Future: Send + 'static,

Start a Network and return a handle to it.

Panics

This method will panic if:

not called from within the context of a tokio runtime.

no private-key or server-name were set.