Struct anemo::Config

#[non_exhaustive]
pub struct Config {

    pub quic: Option<QuicConfig>,
    pub connection_manager_channel_capacity: Option<usize>,
    pub connectivity_check_interval_ms: Option<u64>,
    pub max_connection_backoff_ms: Option<u64>,
    pub connection_backoff_ms: Option<u64>,
    pub connect_timeout_ms: Option<u64>,
    pub max_concurrent_outstanding_connecting_connections: Option<usize>,
    pub max_concurrent_connections: Option<usize>,
    pub peer_event_broadcast_channel_capacity: Option<usize>,
    pub max_frame_size: Option<usize>,
    pub inbound_request_timeout_ms: Option<u64>,
    pub outbound_request_timeout_ms: Option<u64>,
    pub shutdown_idle_timeout_ms: Option<u64>,

}

Configuration for a Network.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

quic: Option<QuicConfig>

Configuration for the underlying QUIC transport.

connection_manager_channel_capacity: Option<usize>

Size of the internal ConnectionManagers mailbox.

One example of how this mailbox is used is for submitting connection requests via Network::connect.

If unspecified, this will default to 128.

connectivity_check_interval_ms: Option<u64>

Trigger connectivity checks every interval.

If unspecified, this will default to 5,000 milliseconds.

max_connection_backoff_ms: Option<u64>

Maximum delay between 2 consecutive attempts to connect with a peer.

If unspecified, this will default to 60,000 milliseconds.

connection_backoff_ms: Option<u64>

The backoff step size, in milliseconds, used to calculate the delay between two consecutive attempts to connect with a peer.

If unspecified, this will default to 10,000 milliseconds.

connect_timeout_ms: Option<u64>

Set a timeout, in milliseconds, for all inbound and outbound connects.

In unspecified, this will default to 10,000 milliseconds.

max_concurrent_outstanding_connecting_connections: Option<usize>

Maximum number of concurrent connections to attempt to establish at a given point in time.

If unspecified, this will default to 100.

max_concurrent_connections: Option<usize>

Maximum number of concurrent connections to have established at a given point in time.

This limit is applied in the following ways:

Inbound connections from KnownPeers with PeerAffinity::High bypass this limit. All other inbound connections are only accepted if the total number of inbound and outbound connections, irrespective of affinity, is less than this limit.

Outbound connections explicitly made by the application via Network::connect or Network::connect_with_peer_id bypass this limit.

Outbound connections made in the background, due to configured KnownPeers, to peers with PeerAffinity::High bypass this limit and are always attempted, while peers with lower affinity respect this limit.

If unspecified, there will be no limit on the number of concurrent connections.

peer_event_broadcast_channel_capacity: Option<usize>

Size of the broadcast channel use for subscribing to PeerEvents via Network::subscribe.

If unspecified, this will default to 128.

max_frame_size: Option<usize>

Set the maximum frame size in bytes.

This controls the maximum size of a request or response.

If unspecified, there will be no limit.

inbound_request_timeout_ms: Option<u64>

Set a timeout, in milliseconds, for all inbound requests.

When an inbound timeout is hit when processing a request a Response is sent to the requestor with a StatusCode::RequestTimeout status code.

In unspecified, no default timeout will be configured.

outbound_request_timeout_ms: Option<u64>

Set a timeout, in milliseconds, for all outbound requests.

When an outbound timeout is hit a timeout error will be returned.

In unspecified, no default timeout will be configured.

shutdown_idle_timeout_ms: Option<u64>

Set a timeout, in milliseconds, until the peers are notified when network is shutting down

If unspecified, then this will default to 1 minute (60 * 1_000 ms)