Struct anemo::Router

pub struct Router { /* private fields */ }

The router type for composing handlers and services.

Implementations

impl Router

pub fn new() -> Self

pub fn route<T>(self, path: &str, service: T) -> Selfwhere
T: Service<Request<Bytes>, Response = Response<Bytes>, Error = Infallible> + Clone + Send + 'static,
T::Future: Send + 'static,

pub fn add_rpc_service<S>(self, service: S) -> Selfwhere
S: Service<Request<Bytes>, Response = Response<Bytes>, Error = Infallible> + RpcService + Clone + Send + 'static,
S::Future: Send + 'static,

pub fn merge<R>(self, other: R) -> Selfwhere
R: Into<Router>,

Merge two routers into one.

This is useful for breaking apps into smaller pieces and combining them into one.

pub fn route_layer<L>(self, layer: L) -> Selfwhere
L: Layer<Route>,
L::Service: Service<Request<Bytes>, Response = Response<Bytes>, Error = Infallible> + Clone + Send + 'static,
<L::Service as Service<Request<Bytes>>>::Future: Send + 'static,

Apply a [tower::Layer] to the router that will only run if the request matches a route.