Enum anemo::types::Address

pub enum Address {
    SocketAddr(SocketAddr),
    HostAndPort {
        host: Box<str>,
        port: u16,
    },
    AddressString(Box<str>),
}

Representation of a network address that is dial-able in Anemo

Variants

SocketAddr(SocketAddr)

A plain SocketAddr

HostAndPort

Fields

host: Box<str>

port: u16

Host and Port where ‘host’ should be either a string representation of an IpAddr address or a host name that will be resolved via DNS.

AddressString(Box<str>)

A string representation of a SocketAddr or a string like <host_name>:<port> pair where <port> is a u16 value.