anemo::types::request
Struct Request

pub struct Request<T> { /* private fields */ }

Implementations

impl<T> Request<T>

pub fn new(body: T) -> Request<T>

pub fn from_parts(parts: RequestHeader, body: T) -> Request<T>

pub fn route(&self) -> &str

pub fn route_mut(&mut self) -> &mut String

pub fn with_route<R: Into<String>>(self, route: R) -> Self

pub fn version(&self) -> Version

pub fn headers(&self) -> &HeaderMap

pub fn headers_mut(&mut self) -> &mut HeaderMap

pub fn with_header<K: Into<String>, V: Into<String>>( self, key: K, value: V, ) -> Self

pub fn extensions(&self) -> &Extensions

pub fn extensions_mut(&mut self) -> &mut Extensions

pub fn with_extension<E: Clone + Send + Sync + 'static>( self, extension: E, ) -> Self

pub fn peer_id(&self) -> Option<&PeerId>

pub fn body(&self) -> &T

pub fn body_mut(&mut self) -> &mut T

pub fn into_body(self) -> T

pub fn inner(&self) -> &T

pub fn inner_mut(&mut self) -> &mut T

pub fn into_inner(self) -> T

pub fn into_parts(self) -> (RequestHeader, T)

pub fn map<F, U>(self, f: F) -> Request<U>
where F: FnOnce(T) -> U,

pub fn set_timeout(&mut self, timeout: Duration)

Set the max duration the request is allowed to take.

The duration will be formatted as nanoseconds as an u64 in the TIMEOUT header.

Example:

use std::time::Duration;
use anemo::Request;

let mut request = Request::new(());

request.set_timeout(Duration::from_secs(30));

let value = request.headers().get("timeout").unwrap();

assert_eq!(
    value,
    &Duration::from_secs(30).as_nanos().to_string(),
);

pub fn with_timeout(self, timeout: Duration) -> Self

Set the max duration the request is allowed to take.

The duration will be formatted as nanoseconds as an u64 in the TIMEOUT header.

Example:

use std::time::Duration;
use anemo::Request;

let request = Request::new(()).with_timeout(Duration::from_secs(30));

let value = request.headers().get("timeout").unwrap();

assert_eq!(
    value,
    &Duration::from_secs(30).as_nanos().to_string(),
);

pub fn timeout(&self) -> Option<Duration>

Returns the previously-set timeout on this Request, or None if not set or invalid.

Example:

use std::time::Duration;
use anemo::Request;

let request = Request::new(()).with_timeout(Duration::from_secs(30));
let timeout = request.timeout().unwrap();

assert_eq!(
    timeout,
    Duration::from_secs(30),
);

impl Request<Bytes>

pub fn empty() -> Self